+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.0] - 2024-07-07
+
 ### Added
 
 - special integration for relay (connection and edge types) and result types (result and success types)
@@ -42,7 +44,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - (No changelog yet)
 
-[Unreleased]: https://github.com/graphinx/graphinx/compare/v0.8.0...HEAD
+[unreleased]: https://github.com/graphinx/graphinx/compare/v0.9.0...HEAD
+[0.9.0]: https://github.com/graphinx/graphinx/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/graphinx/graphinx/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/graphinx/graphinx/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/graphinx/graphinx/releases/tag/v0.6.0